Renaming and small fixes
