+++ mt-kahypar/partition/coarsening/multilevel_coarsener_base.h
@@ -23,7 +23,7 @@
 #include "mt-kahypar/partition/context.h"
 #include "mt-kahypar/partition/refinement/i_refiner.h"
 #include "mt-kahypar/parallel/stl/scalable_vector.h"
-#include "mt-kahypar/partition/coarsening/level.h"
+#include "mt-kahypar/partition/coarsening/coarsening_commons.h"
 
 
 namespace mt_kahypar {
