+++ mt-kahypar/partition/coarsening/multilevel_uncoarsener.cpp
@@ -191,4 +191,4 @@ namespace mt_kahypar {
     }
   }
 
-};
+}
