+++ mt-kahypar/partition/coarsening/multilevel_uncoarsener.h
@@ -21,7 +21,7 @@
 #pragma once
 
 #include "mt-kahypar/partition/context.h"
-#include "mt-kahypar/partition/coarsening/level.h"
+#include "mt-kahypar/partition/coarsening/coarsening_commons.h"
 #include "mt-kahypar/partition/coarsening/i_uncoarsener.h"
 #include "mt-kahypar/partition/coarsening/uncoarsener_base.h"
 namespace mt_kahypar {
