+++ mt-kahypar/partition/coarsening/nlevel_uncoarsener.cpp
@@ -343,4 +343,4 @@ namespace mt_kahypar {
     }
   }
 
-};
+}
