+++ mt-kahypar/partition/coarsening/nlevel_uncoarsener.h
@@ -24,7 +24,7 @@
 #include "mt-kahypar/partition/coarsening/i_uncoarsener.h"
 #include "mt-kahypar/partition/coarsening/uncoarsener_base.h"
 #include "mt-kahypar/partition/refinement/i_refiner.h"
-#include "mt-kahypar/partition/coarsening/level.h"
+#include "mt-kahypar/partition/coarsening/coarsening_commons.h"
 namespace mt_kahypar {
 
   class NLevelUncoarsener : public IUncoarsener,
