+++ mt-kahypar/partition/coarsening/uncoarsener_base.h
@@ -23,7 +23,7 @@
 #include "mt-kahypar/definitions.h"
 #include "mt-kahypar/partition/context.h"
 #include "mt-kahypar/partition/refinement/i_refiner.h"
-#include "mt-kahypar/partition/coarsening/level.h"
+#include "mt-kahypar/partition/coarsening/coarsening_commons.h"
 #include "mt-kahypar/utils/stats.h"
 #include "mt-kahypar/partition/metrics.h"
